refactoring day14
