+++ 2019/src/lib.rs
@@ -13,6 +13,7 @@ pub mod day10;
 pub mod day11;
 pub mod day12;
 pub mod day13;
+pub mod day14;
 pub mod day16;
 pub mod day17;
 pub mod day19;
