+++ 2019/tests/day14_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day14;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(443537, day14::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(2910558, day14::part2());
+    }
+}
